# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/stores": {
      "get": {
        "operationId": "StoresController_findAll",
        "summary": "Get all stores",
        "description": "Retrieve a list of all stores. Public endpoint - no authentication required.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of stores retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Stores"
        ]
      },
      "post": {
        "operationId": "StoresController_create",
        "summary": "Create a new store",
        "description": "Create a new store and get store ID. Public endpoint - used during user signup. Pass store name in body, API will assign a store ID.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoreDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Store created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input"
          }
        },
        "tags": [
          "Stores"
        ]
      }
    },
    "/stores/{id}": {
      "delete": {
        "operationId": "StoresController_remove",
        "summary": "Delete store",
        "description": "Delete a store by ID. ADMIN only. Fails if store has users, inventory, or invoices.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Store UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Store deleted"
          },
          "400": {
            "description": "Cannot delete: store has users, inventory, or invoices"
          },
          "404": {
            "description": "Store not found"
          }
        },
        "tags": [
          "Stores"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/admin/filters": {
      "get": {
        "operationId": "AdminController_getFilters",
        "summary": "Get filter options",
        "description": "Get all available filter options (stores, regions, cities, categories, brands, manufacturers) with their IDs. Use this to get IDs needed for product creation.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Filter options with IDs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stores": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "categories": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "brands": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "manufacturers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/categories": {
      "get": {
        "operationId": "AdminController_getCategories",
        "summary": "Get all categories",
        "description": "Get list of all categories with their IDs. Use these IDs when creating products via POST /products endpoint.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "4158f662-9e15-42ef-9c4e-345b9465693c"
                      },
                      "name": {
                        "type": "string",
                        "example": "Beverages"
                      },
                      "parentId": {
                        "type": "string",
                        "nullable": true,
                        "example": null
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/brands": {
      "get": {
        "operationId": "AdminController_getBrands",
        "summary": "Get all brands",
        "description": "Get list of all brands with their IDs and manufacturer info. Use these IDs when creating products via POST /products endpoint.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of brands",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "4158f662-9e15-42ef-9c4e-345b9465693c"
                      },
                      "name": {
                        "type": "string",
                        "example": "Coca-Cola"
                      },
                      "manufacturerId": {
                        "type": "string",
                        "example": "4158f662-9e15-42ef-9c4e-345b9465693c"
                      },
                      "manufacturerName": {
                        "type": "string",
                        "example": "The Coca-Cola Company"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/manufacturers": {
      "get": {
        "operationId": "AdminController_getManufacturers",
        "summary": "Get all manufacturers",
        "description": "Get list of all manufacturers with their IDs. Use these IDs when creating products via POST /products endpoint.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of manufacturers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "4158f662-9e15-42ef-9c4e-345b9465693c"
                      },
                      "name": {
                        "type": "string",
                        "example": "The Coca-Cola Company"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "summary": "Create product",
        "description": "Create a new product with full details. Requires ADMIN, INVENTORY, or PURCHASE role. \n\n**Note:** This endpoint requires IDs (UUIDs) for category, brand, manufacturer, and store. To get these IDs:\n- `GET /admin/categories` - Get all categories with IDs\n- `GET /admin/brands` - Get all brands with IDs\n- `GET /admin/manufacturers` - Get all manufacturers with IDs\n- `GET /stores` - Get all stores with IDs (public endpoint)\n- `GET /admin/filters` - Get all of the above in one call\n\n**Alternative:** Use `POST /products/admin-create` if you want to create products using names instead of IDs (recommended).\n\nAll IDs must exist in the database. storeId is required - product belongs to a specific store.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "4158f662-9e15-42ef-9c4e-345b9465693c"
                    },
                    "sku": {
                      "type": "string",
                      "example": "PRODUCT-001"
                    },
                    "name": {
                      "type": "string",
                      "example": "Product Name"
                    },
                    "storeId": {
                      "type": "string",
                      "example": "4158f662-9e15-42ef-9c4e-345b9465693c"
                    },
                    "unitPrice": {
                      "type": "number",
                      "example": 50
                    },
                    "unitSizeMl": {
                      "type": "number",
                      "example": 500,
                      "nullable": true,
                      "description": "DEPRECATED: Use unitSizeValue + unitSizeUnit"
                    },
                    "unitSizeValue": {
                      "type": "number",
                      "example": 500,
                      "nullable": true
                    },
                    "unitSizeUnit": {
                      "type": "string",
                      "example": "ML",
                      "enum": [
                        "ML",
                        "L",
                        "G",
                        "KG",
                        "PCS"
                      ],
                      "nullable": true
                    },
                    "category": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    },
                    "brand": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "manufacturer": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "manufacturer": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input (e.g., missing storeId, invalid IDs)"
          },
          "404": {
            "description": "Category, Brand, Manufacturer, or Store not found"
          },
          "409": {
            "description": "Product with this SKU already exists in this store"
          }
        },
        "tags": [
          "Products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "summary": "Get all products",
        "description": "Retrieve a list of products. SALES users see only their store's products. ADMIN/INVENTORY/PURCHASE can filter by storeId query parameter.",
        "parameters": [
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "description": "Store ID (optional for ADMIN/INVENTORY/PURCHASE, auto-set for SALES)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "4158f662-9e15-42ef-9c4e-345b9465693c"
                      },
                      "sku": {
                        "type": "string",
                        "example": "COKE-330ML"
                      },
                      "name": {
                        "type": "string",
                        "example": "Coca-Cola 330ml"
                      },
                      "unitPrice": {
                        "type": "number",
                        "example": 35
                      },
                      "unitSizeMl": {
                        "type": "number",
                        "example": 330,
                        "nullable": true
                      },
                      "category": {
                        "type": "object"
                      },
                      "brand": {
                        "type": "object"
                      },
                      "manufacturer": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/products/admin-create": {
      "post": {
        "operationId": "ProductsController_createFromAdmin",
        "summary": "Create product (admin) - Recommended",
        "description": "Create product with flexible options. **This is the recommended endpoint** because:\n\n- You can use **names** instead of IDs for category, brand, and manufacturer (they will be created if they don't exist)\n- You can use either `categoryId` OR `categoryName` (same for brand/manufacturer)\n- You can optionally add initial stock\n- **storeId is required** - product must belong to a store\n\n**Example:** Use `categoryName: \"Beverages\"` instead of `categoryId: \"uuid-here\"`",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created"
          },
          "400": {
            "description": "Invalid input: e.g. need categoryId or categoryName; storeId when stockQuantity>0"
          }
        },
        "tags": [
          "Products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/products/sku/{sku}": {
      "get": {
        "operationId": "ProductsController_findBySku",
        "summary": "Get product by SKU",
        "description": "Retrieve a product by its SKU (barcode). SALES users see only products from their store. ADMIN/INVENTORY/PURCHASE can use storeId query parameter.",
        "parameters": [
          {
            "name": "sku",
            "required": true,
            "in": "path",
            "description": "Product SKU (barcode)",
            "schema": {
              "example": "COKE-330ML",
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "description": "Store ID (optional for ADMIN/INVENTORY/PURCHASE, auto-set for SALES)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": [
          "Products"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Meri Dukaan POS + Admin Analytics API",
    "description": "Complete API documentation for Meri Dukaan Point of Sale system with Admin Analytics",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "StoreResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "4158f662-9e15-42ef-9c4e-345b9465693c",
            "description": "Store UUID"
          },
          "name": {
            "type": "string",
            "example": "Store Karachi Central",
            "description": "Store name"
          },
          "region": {
            "type": "string",
            "example": "Sindh",
            "description": "Store region"
          },
          "city": {
            "type": "string",
            "example": "Karachi",
            "description": "Store city"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-15T10:00:00.000Z",
            "description": "Store creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-15T10:00:00.000Z",
            "description": "Store last update timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "region",
          "city",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateStoreDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My Store Name",
            "description": "Store name"
          },
          "region": {
            "type": "string",
            "example": "Punjab",
            "description": "Store region (optional, defaults to empty string)"
          },
          "city": {
            "type": "string",
            "example": "Lahore",
            "description": "Store city (optional, defaults to empty string)"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}